check che right thing to prevent the crash in the timeout thread
Though this seems to be a dead end, the lines CG(timed_out) = 0; looks
like will never be executed in the TS build. Or at least it'll be executed
by luck when the scheduler has decided to let the master thread run for
longer. NTS should be fine with that however.