Revert "Fix #71152: mt_rand() returns the different values from original mt19937ar.c"
This reverts commit b816893efbb631ef4b171e2c38372dc98bac3c56.

`mt_rand()` is seedable with `mt_srand()` which means it can be used to (re)produce specific streams of numbers. All code (no matter how few instances that may be) that previously depended on this behaviour will no longer produce the same results.

This kind of change needs to be discussed before being committed.